requirements.txt
requests  
pandas  
numpy  
ccxt  # For connecting to crypto exchanges  
pytest  # For testing  
python-dotenv  # For managing environment variables  
python-telegram-bot  # For sending messages to Telegram
SQLAlchemy  # For connecting to databases
mysql-connector-python
pandas_ta  # For technical analysis